+++ packages/components/src/Chip/Chip.tsx
@@ -135,7 +135,6 @@ function getTooltipMessage(
 
 ChipComponent.displayName = "Chip";
 
-// Required to assign sub-components to the Chip component when it is a forwardRef
 const ChipNamespace = Object.assign(ChipComponent, {
   Prefix: ChipPrefix,
   Suffix: ChipSuffix,